-

ABC — Module.

ABC Module providing types (abc_pmc_plan_type, abc_population_type) and methods (generate_theta, init_abc, update_abc_pop, run_abc) for using Approximate Bayesian Computing

source

abc_pmc_plan_type(gd::Function,css::Function,cd::Function,p)

ABC.abc_population_type — Method.

abc_population_type(num_param::Integer, num_particles::Integer; accept_log::abc_log_type = abc_log_type(), reject_log::abc_log_type = abc_log_type(), repeats::Array{Int64,1} = zeros(Int64,num_particles) )

num_param: Number of model parameters for generating simulated data num_particles: Number of particles for sequential importance sampler

Optional parameters: accept_log: Log of accepted parameters/summary statistics/distances reject_log: Log of rejected parameters/summary statistics/distances repeats: Array indicating which particles have been repeated from previous generation

source